central/app/requirements.txt
# .env file
python-dotenv>=1.0.0

# Core scientific stack
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.2.0
joblib>=1.3.0

# Machine Learning (model serving)
tensorflow>=2.15.0

# API
fastapi>=0.110.0
uvicorn>=0.29.0

# Optional (for CSV serving)
python-multipart>=0.0.6
